+++ tests/integration/test_logtest/test_remove_old_sessions/test_remove_old_session_for_inactivity.py
@@ -56,6 +56,7 @@ from logtest import callback_remove_session, callback_session_initialized
 from wazuh_testing.constants.paths.sockets import LOGTEST_SOCKET_PATH
 from wazuh_testing.global_parameters import GlobalParameters
 from wazuh_testing.constants.paths.logs import WAZUH_LOG_PATH
+from wazuh_testing.modules.analysisd.configuration import ANALYSISD_DEBUG
 from wazuh_testing.tools.monitors import file_monitor
 from wazuh_testing.utils import configuration
 
@@ -73,7 +74,7 @@ t_configurations = configuration.load_configuration_template(t_config_path, t_co
 # Variables
 receiver_sockets_params = [(LOGTEST_SOCKET_PATH, 'AF_UNIX', 'TCP')]
 receiver_sockets = None
-local_internal_options = {'analysisd.debug': '1'}
+local_internal_options = {ANALYSISD_DEBUG: '1'}
 create_session_data = {'version': 1, 'command': 'log_processing',
                        'parameters': {'event': 'Oct 15 21:07:56 linux-agent sshd[29205]: Invalid user blimey '
                                       'from 18.18.18.18 port 48928',
